\name{banana}
\alias{banana}
\docType{data}
\title{
Banana Adsorption study
}
\description{
Adsorption of Copper(II) using Musa acuminata x balbisiana seeds as bioadsorbent
}
\usage{data("banana")}
\format{
  A data frame with 6 observations on the following 2 variables.
  \describe{
    \item{Ce}{a numeric vector for the equilibrium concentration}
    \item{Qe}{a numeric vector for the adsorbed concentration}
  }
}
\source{
Development of an R package for chemical adsorption isotherms, kinetics and thermodynamics(2019), BS Chemistry thesis , Polytechnic University of the Philippines
}
\keyword{datasets}
