% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bauduanalysis.R
\name{bauduanalysis}
\alias{bauduanalysis}
\title{Baudu Isotherm Analysis}
\usage{
bauduanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Baudu isotherm
}
\description{
Reduced form of Langmuir Isotherm
}
\examples{
bauduanalysis(c(1,2,3,4,5),c(1,2,3,4,5))

}
