% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bauduplot.R
\name{bauduplot}
\alias{bauduplot}
\title{Baudu Isotherm Plot}
\usage{
bauduplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot of Baudu isotherm analysis
 @export
}
\description{
Plot of the analysis of Baudu Isotherm
}
