% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betanalysis.r
\name{BETanalysis}
\alias{BETanalysis}
\title{BET Isotherm Analysis}
\usage{
BETanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilbrium concentration}

\item{Qe}{the numerical value for the adsorbed concentration}
}
\value{
the Linear model for the BET isotherm analysis
}
\description{
An isotherm that takes account of the possibility that the monolayer in the Langmuir adsorption isotherm can act as a substrate for further adsorption.
}
