% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETplot.R
\name{betplot-deprecated}
\alias{betplot-deprecated}
\title{BET Isotherm Plot-deprecated}
\usage{
betplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the BET adsorption isoth
}
\description{
Plot of the analysis of BET Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
