% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETplot.R
\name{betplot}
\alias{betplot}
\title{BET Isotherm Plot}
\usage{
betplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the BET adsorption isotherm
}
\description{
Plot of the analysis of BET Isotherm
}
\examples{
betplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
