% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaG.R
\name{deltaG}
\alias{deltaG}
\title{Gibbs free energy of Adsorption}
\usage{
deltaG(t, K)
}
\arguments{
\item{t}{temperature used in the experiment}

\item{K}{equilirium constant for the adsorption process}
}
\value{
the Gibbs free energy value for the given adsorption process in terms of Kj(kilojoule) per mole
}
\description{
Defines the spontaneity of an adsoprtion process base from the value. If it is negative is is spontaneous and if the result is positive,the reaction is not spontaneous.
}
\examples{
deltaG(25,1234)
}
