% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubininradushkevichplot.R
\name{drplot-deprecated}
\alias{drplot-deprecated}
\title{Dubinin-Radushkevich Isotherm Plot-deprecated}
\usage{
drplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot of Dubinin-Radushkevich isotherm analysis
}
\description{
Plot of the analysis of Dubinun-Radushkevich Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
