% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubininradushkevichplot.R
\name{drplot}
\alias{drplot}
\title{Dubinin-Radushkevich Isotherm Plot}
\usage{
drplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot of Dubinin-Radushkevich isotherm analysis
}
\description{
Plot of the analysis of Dubinun-Radushkevich Isotherm
}
\examples{
drplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
