% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dubininradushkevichanalysis.r
\name{dubininradanalysis}
\alias{dubininradanalysis}
\title{Dubinin Radushkevich Isotherm Analysis}
\usage{
dubininradanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Dubinin Radushkevich isotherm analysis
}
\description{
It is only suitable for intermediate range of adsorbate concentrations because it exhibits unrealistic asymptotic behavior and does not predict Henry’s laws at low pressure.
}
\examples{
dubininradanalysis (c(1,2,3,4,5),c(1,2,3,4,5))
}
