% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elovichkinetics.R
\name{elovichkinetics}
\alias{elovichkinetics}
\title{Elovich Kinetics}
\usage{
elovichkinetics(t, Qt)
}
\arguments{
\item{t}{duration for the experiment}

\item{Qt}{the numerical value for the concentration at given time}
}
\value{
The regression analysis for the Elovich Kinetics
}
\description{
This equation assumes that the actual solid surfaces are energetically heterogenous and that neither desorption nor interactions betwen adsorbed species could subcutaneously affect the kinetics of adsorption at low surface coverage. (Mercado-Borayo, et. al, 2014)
}
