% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elovichplot.R
\name{elovichplot-deprecated}
\alias{elovichplot-deprecated}
\title{Elovich Isotherm Plot-deprecated}
\usage{
elovichplot(Qe, Ce)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Elovich isotherm analysis
}
\description{
Plot of the analysis of Elovich Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
