% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elovichplot.R
\name{elovichplot}
\alias{elovichplot}
\title{Elovich Isotherm Plot}
\usage{
elovichplot(Qe, Ce)
}
\arguments{
\item{Qe}{the numerical value for the adsorbed capacity}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the LSRL plot for Elovich isotherm analysis
}
\description{
Plot of the analysis of Elovich Isotherm
}
\examples{
elovichplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
