% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fowlerguggenheimplot.R
\name{fgplot-deprecated}
\alias{fgplot-deprecated}
\title{Fowler Guggenheim Isotherm Plot-deprecated}
\usage{
fgplot(theta, Ce)
}
\arguments{
\item{theta}{the numerical value for the fractional coverage}

\item{Ce}{the numerical value for the equilibrium concentration}
}
\value{
the LSRL plot for Flory Huggins isotherm analysis
}
\description{
Plot of the analysis of Fowler Guggenheim Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
