% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fowlerguggenheimplot.R
\name{fgplot}
\alias{fgplot}
\title{Fowler Guggenheim Isotherm Plot}
\usage{
fgplot(theta, Ce)
}
\arguments{
\item{theta}{the numerical value for the fractional coverage}

\item{Ce}{the numerical value for the equilibrium concentration}
}
\value{
the LSRL plot for Flory Huggins isotherm analysis
}
\description{
Plot of the analysis of Fowler Guggenheim Isotherm
}
\examples{
fgplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
