% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floryhugginsanalysis.R
\name{fhanalysis}
\alias{fhanalysis}
\title{Flory Huggins Isotherm Analysis}
\usage{
fhanalysis(theta, Co)
}
\arguments{
\item{theta}{the numeric value for the fractional coverage}

\item{Co}{the numeric value for the initial concentration}
}
\value{
the regression analysis for the Flory Huggins Isotherm
}
\description{
Flory Huggins Isotherm Analysis
}
