% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floryhugginsplot.R
\name{fhplot}
\alias{fhplot}
\title{Flory Huggins Isotherm Plot}
\usage{
fhplot(Ce, theta)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{theta}{the numerical value for the fractional coverage}
}
\value{
the LSRL plot for Flory Huggins isotherm analysis
}
\description{
Plot of the analysis of Flory Huggins
}
