% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1storderkinetics.R
\name{firstorder}
\alias{firstorder}
\title{Pseudo-1st Order Kinetics}
\usage{
firstorder(t, Ce)
}
\arguments{
\item{t}{duration of the experiment}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the regression analysis for the first order kinetics
}
\description{
A first-order rate equation which is believed to be the earliest model that was presented by Lagergen (1898) to describe the kinetic process of liqud-solid phase adsorption of oxalic acid and malonic acid onto charcoal pertains to the adsorption rate based on the adsorption capacity.
}
\examples{
firstorder(c(1,2,3,4,5),c(1,2,3,4,5))
}
