% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5paramplot.R
\name{fivePplot-deprecated}
\alias{fivePplot-deprecated}
\title{Five Parameter Isotherm Plot-deprecated}
\usage{
fivePplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Five parameter isotherm analysis
}
\description{
Plot of the analysis of Five Parameter Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
