% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5paramplot.R
\name{fivePplot}
\alias{fivePplot}
\title{Five Parameter Isotherm Plot}
\usage{
fivePplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Five parameter isotherm analysis
}
\description{
Plot of the analysis of Five Parameter Isotherm
}
\examples{
fivePplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
