% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractionalpower.R
\name{fractionalpower}
\alias{fractionalpower}
\title{Fractional Power Kinetic Model}
\usage{
fractionalpower(t, qt)
}
\arguments{
\item{t}{duration of the experiment}

\item{qt}{the numerical value for the equilibrium capacity}
}
\value{
the regression analysis for the first order kinetics
}
\description{
Calculates Kt parameter of the Fractional power kinetic model.
}
\examples{
fractionalpower (c(1,2,3,4,5),c(1,2,3,4,5))
}
