% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freundlichplot.R
\name{freundlichplot-deprecated}
\alias{freundlichplot-deprecated}
\title{Freundlich Isotherm plot-deprecated}
\usage{
freundlichplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the freundlich adsorption isotherm
}
\description{
Freundlich Isotherm plot-deprecated
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
