% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freundlichplot.R
\name{freundlichplot}
\alias{freundlichplot}
\title{Title}
\usage{
freundlichplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the freundlich adsorption isotherm
}
\description{
Title
}
\examples{
freundlichplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
