% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritzschlunderanalysis.r
\name{fritzanalysis}
\alias{fritzanalysis}
\title{Fritz Schlunder Isotherm Analysis}
\usage{
fritzanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Fritz Schlunder isotherm analysis
}
\description{
An empirical equation which can fit a wide range of experimental results because of the large number of coefficients in the isotherm.
}
\examples{
fritzanalysis(moringa$Ce,moringa$Qe)
}
