% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritzschlunderplot.r
\name{fsplot-deprecated}
\alias{fsplot-deprecated}
\title{Fritz Schlunder Isotherm Plot-deprecated}
\usage{
fsplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot fot Fritz Schlunder isotherm analysis
}
\description{
Plot of the analysis of Fritz Schlunder Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
