% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritzschlunderplot.r
\name{fsplot}
\alias{fsplot}
\title{Fritz Schlunder Isotherm Plot}
\usage{
fsplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot fot Fritz Schlunder isotherm analysis
}
\description{
Plot of the analysis of Fritz Schlunder Isotherm
}
\examples{
fsplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
