% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halseyplot.R
\name{halseyplot-deprecated}
\alias{halseyplot-deprecated}
\title{Halsey Isotherm Plot-deprecated}
\usage{
halseyplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Halsey isotherm analysis
}
\description{
Plot of the analysis of Halsey Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
