% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halseyplot.R
\name{halseyplot}
\alias{halseyplot}
\title{Halsey Isotherm Plot}
\usage{
halseyplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Halsey isotherm analysis
}
\description{
Plot of the analysis of Halsey Isotherm
}
\examples{
halseyplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
