% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilldeboerplot.R
\name{hdplot-deprecated}
\alias{hdplot-deprecated}
\title{Hill Deboer Isotherm-deprecated}
\usage{
hdplot(theta, Ce)
}
\arguments{
\item{theta}{a numeric vector consists of fractional coverage}

\item{Ce}{a numeric vector consists of equilibrium concentration}
}
\value{
the LSRL plot for Hill Deboer isotherm
}
\description{
Plot of the analysis of Hill Deboer Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
