% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilldeboerplot.R
\name{hdplot}
\alias{hdplot}
\title{Hill Deboer Isotherm}
\usage{
hdplot(theta, Ce)
}
\arguments{
\item{theta}{a numeric vector consists of fractional coverage}

\item{Ce}{a numeric vector consists of equilibrium concentration}
}
\value{
the LSRL plot for Hill Deboer isotherm
}
\description{
Plot of the analysis of Hill Deboer Isotherm
}
\examples{
hdplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
