% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henryanalysis.R
\name{henryanalysis}
\alias{henryanalysis}
\title{Henry Isotherm}
\usage{
henryanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Henry isotherm analysis
}
\description{
It describes an appropriate fit to the adsorption of adsorbate at relatively low concentrations such that all adsorbate molecules are secluded from their nearest neighbours.
}
\examples{
henryanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
