% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henryplot.R
\name{henryanalysisplot-deprecated}
\alias{henryanalysisplot-deprecated}
\title{Henry Isotherm Plot-deprecated}
\usage{
henryanalysisplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Henry isotherm analysis
}
\description{
Plot of the analysis of Henry Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
