% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillanalysis.R
\name{hillanalysis}
\alias{hillanalysis}
\title{Hill Isotherm}
\usage{
hillanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Hill isotherm analysis
}
\description{
Hill Isotherm
}
\examples{
hillanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
