% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillplot.R
\name{hillplot-deprecated}
\alias{hillplot-deprecated}
\title{Hill Isotherm Plot-deprecated}
\usage{
hillplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Hill isotherm analysis
}
\description{
Plot of the analysis of Hill Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
