% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillplot.R
\name{hillplot}
\alias{hillplot}
\title{Hill Isotherm Plot}
\usage{
hillplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Hill isotherm analysis
}
\description{
Plot of the analysis of Hill Isotherm
}
\examples{
hillplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
