% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harkinsjuraplot.R
\name{hjplot-deprecated}
\alias{hjplot-deprecated}
\title{Harkins- Jura Isotherm Plot-deprecated}
\usage{
hjplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for a equilbrium concentration}

\item{Qe}{the numerical value for the adsorbed quantities}
}
\value{
the LSRL plot for the Harkins Jura Isotherm
}
\description{
Plot of the analysis of Harkins-Jura Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
