% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harkinsjuraplot.R
\name{hjplot}
\alias{hjplot}
\title{Harkins- Jura Isotherm Plot}
\usage{
hjplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for a equilbrium concentration}

\item{Qe}{the numerical value for the adsorbed quantities}
}
\value{
the LSRL plot for the Harkins Jura Isotherm
}
\description{
Plot of the analysis of Harkins-Jura Isotherm
}
\examples{
hjplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
