% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jossensanalysis.R
\name{jossensanalysis}
\alias{jossensanalysis}
\title{Jossens Isotherm}
\usage{
jossensanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Jossens isotherm analysis
}
\description{
It predicts a simple equation based on the energy distribution of adsorbate-adsorbent interactions at adsorption sites.This model assumes that the adsorbent has heterogeneous surface with respect to the interactions it has with the adsorbate.
}
\examples{
jossensanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
