% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jossensplot.R
\name{jossensplot-deprecated}
\alias{jossensplot-deprecated}
\title{Jossens Isotherm Plot-deprecated}
\usage{
jossensplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Jossen's isotherm
}
\description{
Plot of the analysis of Jossens Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
