% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jovanovicanalysis.R
\name{jovanovicanalysis}
\alias{jovanovicanalysis}
\title{Jovanovic Isotherm}
\usage{
jovanovicanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Jovanovic isotherm analysis
}
\description{
It is predicated on the assumptions contained in the Langmuir model, but in addition the possibility of some mechanical contacts between the adsorbate and adsorbent.
}
\examples{
jovanovicanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
