% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jovanovicplot.R
\name{jovanovicplot-deprecated}
\alias{jovanovicplot-deprecated}
\title{Jovanovic Isotherm Plot-deprecated}
\usage{
jovanovicplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Jovanovic isotherm analysis
}
\description{
Plot of the analysis of Jovanovic Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
