% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kahnanalysis.R
\name{kahnanalysis}
\alias{kahnanalysis}
\title{Kahn Isotherm}
\usage{
kahnanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Kahn isotherm analysis
}
\description{
Kahn Isotherm
}
\examples{
kahnanalysis(moringa$Ce,moringa$Qe)
}
