% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kahnplot.R
\name{kahnplot-deprecated}
\alias{kahnplot-deprecated}
\title{Kahn Isotherm plot-deprecated}
\usage{
kahnplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the LSRL plot for Kahn isotherm analysis
}
\description{
Plot of the analysis of Kahn Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
