% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koblecarriganplot.R
\name{kcarriganplot-deprecated}
\alias{kcarriganplot-deprecated}
\title{Koble-Carrigan Isotherm Plot-deprecated}
\usage{
kcarriganplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for koble carrigan isotherm analysis
}
\description{
Plot of the analysis of Koble-Carrigan Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
