% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koblecarriganplot.R
\name{kcarriganplot}
\alias{kcarriganplot}
\title{Koble-Carrigan Isotherm Plot}
\usage{
kcarriganplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for koble carrigan isotherm analysis
}
\description{
Plot of the analysis of Koble-Carrigan Isotherm
}
\examples{
kcarriganplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
