% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiselevanalysis.R
\name{kiselevanalysis}
\alias{kiselevanalysis}
\title{Kiselev Isotherm Analysis}
\usage{
kiselevanalysis(Kh, Ce)
}
\arguments{
\item{Kh}{the numerical value representing Kiselev analysis}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the linear regression and the parameters for the Kiselev isotherm analysis
}
\description{
Kiselev Isotherm Analysis
}
\examples{
kiselevanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
