% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiselevplot.R
\name{kiselevplot-deprecated}
\alias{kiselevplot-deprecated}
\title{Kiselev Isotherm Plot-deprecated}
\usage{
kiselevplot(Qe, Ce)
}
\arguments{
\item{Kh}{the numerical value for the kiselev constant}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the plot for kiselev isotherm analysis
}
\description{
Plot of the analysis of Kiselev Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
