% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiselevplot.R
\name{kiselevplot}
\alias{kiselevplot}
\title{Kiselev Isotherm Plot}
\usage{
kiselevplot(Kh, Ce)
}
\arguments{
\item{Kh}{the numerical value for the kiselev constant}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the plot for kiselev isotherm analysis
}
\description{
Plot of the analysis of Kiselev Isotherm
}
\examples{
kiselevplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
