% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koblecarrigananalysis.R
\name{koblecarrigananalysis}
\alias{koblecarrigananalysis}
\title{Koble Carrigan Isotherm Analysis}
\usage{
koblecarrigananalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Koble Carrigan isotherm analysis
}
\description{
Koble Carrigan Isotherm Analysis
}
\examples{
koblecarrigananalysis(moringa$Ce,moringa$Qe)
}
