% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langmuiranalysis.R
\name{langmuiranalysis}
\alias{langmuiranalysis}
\title{Langmuir Isotherm Analysis}
\usage{
langmuiranalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the regression analysis for the langmuir isotherm analysis
}
\description{
It was primarily designed to describe gas-solid phase adsorption is also used to quantify and contrast the adsorptive capacity of various adsorbents [12]. Langmuir isotherm accounts for the surface coverage by balancing the relative rates of adsorption and desorption (dynamic equilibrium).
}
\examples{
langmuiranalysis(moringa$Ce,moringa$Qe)
}
