% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langmuirFplot.R
\name{langmuirFplot-deprecated}
\alias{langmuirFplot-deprecated}
\title{Langmuir-Freundlich Isotherm Plot-deprecated}
\usage{
langmuirFplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilbrium concentration}

\item{Qe}{the numerical value for the adsorbed concentration}
}
\value{
the LSRL for the Langmuir-Freundlich isotherm analysis
}
\description{
Plot of the analysis of Langmuir-FreundLich Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
