% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langmuirplot.R
\name{langmuirplot}
\alias{langmuirplot}
\title{Langmuir Plot}
\usage{
langmuirplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the langmuir adsorption isotherm
}
\description{
Langmuir Plot
}
\examples{
langmuirplot(moringa$Ce,moringa$Qe)
}
