% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marckzewskijaroniecplot.R
\name{marcJplot}
\alias{marcJplot}
\title{Marckzewski Jaroniec Isotherm Plot}
\usage{
marcJplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the Marckzewski Jaroniec isotherm analysis
}
\description{
Plot of the analysis of Marckzewski Jaroniec Isotherm
}
\examples{
marcJplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
