% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marckzewskijaroniecanalysis.R
\name{mjanalysis}
\alias{mjanalysis}
\title{Marckzewski Jaroniec Isotherm Analysis}
\usage{
mjanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Marckzewski Jaroniec isotherm analysis
}
\description{
Marckzewski Jaroniec Isotherm Analysis
}
