% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moringa.r
\docType{data}
\name{moringa}
\alias{moringa}
\title{Moringa Adsorption Study}
\format{A data frame with 6 observations on the following 2 variables.
\describe{
\item{\code{Ce}}{a numeric vector}
\item{\code{Qe}}{a numeric vector}

}
...}
\usage{
moringa
}
\description{
Actual adosrption experiment results using Moringa oleifera seeds.
}
\keyword{datasets}
