% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5paramplot.R, R/BETplot.R,
%   R/PUPAIM-deprecated.r, R/WVVplot.R, R/dubininradushkevichplot.R,
%   R/elovichplot.R, R/floryhugginsplot.R, R/fowlerguggenheimplot.R,
%   R/freundlichplot.R, R/fritzschlunderplot.r, R/halseyplot.R,
%   R/harkinsjuraplot.R, R/henryplot.R, R/hilldeboerplot.R,
%   R/hillplot.R, R/jossensplot.R, R/jovanovicplot.R, R/kahnplot.R,
%   R/kiselevplot.R, R/koblecarriganplot.R, R/langmuirFplot.R,
%   R/langmuirplot.R, R/marckzewskijaroniecplot.R, R/raudkeprausniizplot.R,
%   R/redlichpetersonplot.R, R/sipsplot.R, R/summaryplot.R, R/temkinplot.R,
%   R/tothplot.R
\name{fivePplot}
\alias{fivePplot}
\alias{betplot}
\alias{PUPAIM-deprecated}
\alias{WVVplot}
\alias{drplot}
\alias{elovichplot}
\alias{fhplot}
\alias{fgplot}
\alias{freundlichplot}
\alias{fsplot}
\alias{halseyplot}
\alias{hjplot}
\alias{henryanalysisplot}
\alias{hdplot}
\alias{hillplot}
\alias{jossensplot}
\alias{jovanovicplot}
\alias{kahnplot}
\alias{kiselevplot}
\alias{kcarriganplot}
\alias{langmuirFplot}
\alias{langmuirplot}
\alias{marcJplot}
\alias{raudkePplot}
\alias{rpplot}
\alias{sipsplot}
\alias{summaryplots}
\alias{temkinplot}
\alias{tothplot}
\title{Deprecated functions in package \pkg{PUPAIM}.}
\usage{
fivePplot(Ce, Qe)

betplot(Ce, Qe)

WVVplot(Qe, Ce)

drplot(Ce, Qe)

elovichplot(Qe, Ce)

fhplot(Ce, theta)

fgplot(theta, Ce)

freundlichplot(Ce, Qe)

fsplot(Ce, Qe)

halseyplot(Ce, Qe)

hjplot(Ce, Qe)

henryanalysisplot(Ce, Qe)

hdplot(theta, Ce)

hillplot(Ce, Qe)

jossensplot(Ce, Qe)

jovanovicplot(Ce, Qe)

kahnplot(Ce, Qe)

kiselevplot(Qe, Ce)

kcarriganplot(Ce, Qe)

langmuirFplot(Ce, Qe)

langmuirplot(Ce, Qe)

marcJplot(Ce, Qe)

raudkePplot(Ce, Qe)

rpplot(Ce, Qe)

sipsplot(Qe, Ce)

summaryplots(Ce, Qe)

temkinplot(Ce, Qe)

tothplot(Ce, theta)
}
\description{
The functions listed below are deprecated and will be defunct in
  the near future. When possible, alternative functions with similar
  functionality are also mentioned. Help pages for deprecated functions are
  available at \code{help("-deprecated")}.
}
\section{\code{fivePplot}}{

For \code{fivePplot}, use \code{\link{fiveparamanalysis}}.
}

\section{\code{betplot}}{

For \code{betplot}, use \code{\link{BETanalysis}}.
}

\section{\code{WVVplot}}{

For \code{WVVplot}, use \code{\link{webervvanalysis}}.
}

\section{\code{drplot}}{

For \code{drplot}, use \code{\link{dubininradanalysis}}.
}

\section{\code{elovichplot}}{

For \code{elovichplot}, use \code{\link{elovichanalysis}}.
}

\section{\code{fhplot}}{

For \code{fhplot}, use \code{\link{fhanalysis}}.
}

\section{\code{fgplot}}{

For \code{fgplot}, use \code{\link{fowlerganalysis}}.
}

\section{\code{freundlichplot}}{

For \code{freundlichplot}, use \code{\link{freundlichanalysis}}.
}

\section{\code{fsplot}}{

For \code{fsplot}, use \code{\link{fritzanalysis}}.
}

\section{\code{halseyplot}}{

For \code{halseyplot}, use \code{\link{halseyanalysis}}.
}

\section{\code{hjplot}}{

For \code{hjplot}, use \code{\link{harkinsjuraanalysis}}.
}

\section{\code{henryanalysisplot}}{

For \code{henryanalysisplot}, use \code{\link{henryanalysis}}.
}

\section{\code{hdplot}}{

For \code{hdplot}, use \code{\link{hilldeboeranalysis}}.
}

\section{\code{hillplot}}{

For \code{hillplot}, use \code{\link{hillanalysis}}.
}

\section{\code{jossensplot}}{

For \code{jossensplot}, use \code{\link{jossensanalysis}}.
}

\section{\code{jovanovicplot}}{

For \code{jovanovicplot}, use \code{\link{jovanovicanalysis}}.
}

\section{\code{kahnplot}}{

For \code{kahnplot}, use \code{\link{kahnanalysis}}.
}

\section{\code{kiselevplot}}{

For \code{kiselevplot}, use \code{\link{kiselevanalysis}}.
}

\section{\code{kcarriganplot}}{

For \code{kcarriganplot}, use \code{\link{koblecarrigananalysis}}.
}

\section{\code{langmuirFplot}}{

For \code{langmuirFplot}, use \code{\link{langmuirFanalysis}}.
}

\section{\code{langmuirplot}}{

For \code{langmuirplot}, use \code{\link{langmuiranalysis}}.
}

\section{\code{marcJplot}}{

For \code{marcJplot}, use \code{\link{mjanalysis}}.
}

\section{\code{raudkePplot}}{

For \code{raudkePplot}, use \code{\link{raudkepanalysis}}.
}

\section{\code{rpplot}}{

For \code{rpplot}, use \code{\link{redlichpanalysis}}.
}

\section{\code{sipsplot}}{

For \code{sipsplot}, use \code{\link{sipsanalysis}}.
}

\section{\code{summaryplots}}{

For \code{summaryplots}, use \code{\link{summaryanalysis}}.
}

\section{\code{temkinplot}}{

For \code{temkinplot }, use \code{\link{temkinanalysis}}.
}

\section{\code{tothplot}}{

For \code{tothplot}, use \code{\link{tothanalysis}}.
}

\keyword{internal}
