% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUPAIM.R
\docType{package}
\name{PUPAIM}
\alias{PUPAIM}
\title{PUPAIM Package}
\description{
The PUPAIM R package can generally fit any adsorption experimental data to any of the 52 available adsorption isotherm models - 29 nonlinear models and 23 linear models. This package provides parameter estimation, model accuracy analysis, model error analysis, and adsorption plot created using the package ggplot2. This package will help the users for a much easier way of adsorption model data fitting.
}
\author{
Keith Ostan \email{ktostan@iskolarngbayan.pup.edu.ph}

Paul Angelo Manlapaz \email{pacmanlapaz@iskolarngbayan.pup.edu.ph}

Jemimah Christine Mesias \email{jclmesias@iskolarngbayan.pup.edu.ph}

Chester Deocaris \email{ccdeocaris@pup.edu.ph}
}
