% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raudkeprausniizplot.R
\name{radkePplot}
\alias{radkePplot}
\title{Raudke Prausniiz Isotherm Plot}
\usage{
radkePplot(Ce, Qe)
}
\arguments{
\item{Ce}{Ce the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
Provides the plot for Raudke Prausniiz isotherm model
}
\description{
Plot of the analysis of Raudke Prausniiz Isotherm
}
\examples{
radkePplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
