% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raudkeprausniizplot.R
\name{raudkePplot-deprecated}
\alias{raudkePplot-deprecated}
\title{Raudke Prausniiz Isotherm Plot-deprecated}
\usage{
raudkePplot(Ce, Qe)
}
\arguments{
\item{Ce}{Ce the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
Provides the plot for Raudke Prausniiz isotherm model
}
\description{
Plot of the analysis of Raudke Prausniiz Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
