% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlichpetersonanalysis.r
\name{redlichpanalysis}
\alias{redlichpanalysis}
\title{Redlich Peterson Isotherm Analysis}
\usage{
redlichpanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Redlich Peterson isotherm analysis
}
\description{
Redlich Peterson Isotherm Analysis
}
\examples{
redlichpanalysis(moringa$Ce,moringa$Qe)
}
