% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlichpetersonplot.R
\name{rpplot-deprecated}
\alias{rpplot-deprecated}
\title{Redlich Peterson Isotherm Plot-deprecated}
\usage{
rpplot(Ce, Qe)
}
\arguments{
\item{Ce}{Ce the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
Provides the plot for Redlich Peterson isotherm model
}
\description{
Plot of the analysis of Redlich Peterson Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
