% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlichpetersonplot.R
\name{rpplot}
\alias{rpplot}
\title{Redlich Peterson Isotherm Plot}
\usage{
rpplot(Ce, Qe)
}
\arguments{
\item{Ce}{Ce the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
Provides the plot for Redlich Peterson isotherm model
}
\description{
Plot of the analysis of Redlich Peterson Isotherm
}
\examples{
rpplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
