% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2ndorderkinetics.R
\name{secondorder}
\alias{secondorder}
\title{Pseudo-2nd Order Kinetics}
\usage{
secondorder(t, Ce)
}
\arguments{
\item{t}{duration of the experiment}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the regression analysis for the second order kinetics
}
\description{
The pseudo-second order model describes the adsorption reaction rate with depndent energetically hetrogenousites on the adsorbent. (Mercado-Borayo, et. al., 2014)
}
\examples{
secondorder(c(1,2,3,4,5),c(1,2,3,4,5))
}
