% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipsanalysis.R
\name{sipsanalysis}
\alias{sipsanalysis}
\title{Sips Isotherm Analysis}
\usage{
sipsanalysis(Qe, Ce)
}
\arguments{
\item{Qe}{the numerical value for the adsorbed concentration}

\item{Ce}{the numerical value for the equilbrium concentration}
}
\value{
the Linear model for the Sips isotherm analysis
}
\description{
Sips Isotherm Analysis
}
\examples{
sipsanalysis(c(1,2,3,4,5),c(1,2,3,4,5))

}
