% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipsplot.R
\name{sipsplot-deprecated}
\alias{sipsplot-deprecated}
\title{Sips Isotherm Plot-deprecated}
\usage{
sipsplot(Qe, Ce)
}
\arguments{
\item{Ce}{the numerical value for the equilbrium concentration}

\item{Qe}{the numerical value for the adsorbed concentration}
}
\value{
the LSRL of Sips analysis
}
\description{
Plot of the analysis of Sips Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
