% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipsplot.R
\name{sipsplot}
\alias{sipsplot}
\title{Sips Isotherm Plot}
\usage{
sipsplot(Qe, Ce)
}
\arguments{
\item{Qe}{the numerical value for the adsorbed concentration}

\item{Ce}{the numerical value for the equilbrium concentration}
}
\value{
the LSRL of Sips analysis
}
\description{
Plot of the analysis of Sips Isotherm
}
\examples{
sipsplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
