% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryanalysis.R
\name{summaryanalysis}
\alias{summaryanalysis}
\title{Summary of the Isotherm Analysis}
\usage{
summaryanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
summary of the nonlinear and linear fitting for different adsorption isotherm models
}
\description{
1-2 sentences Summarize the analysis for different isotherm models
}
