% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryplot.R
\name{summaryplots-deprecated}
\alias{summaryplots-deprecated}
\title{Summary of Plots of the Isotherm Analysis-deprecated}
\usage{
summaryplots(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the summary of the LSRL plot for different adsorption isotherm
}
\description{
Summarize the plot for different isotherm models
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
