% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temkinanalysis.R
\name{temkinanalysis}
\alias{temkinanalysis}
\title{Temkin Isotherm Analysis}
\usage{
temkinanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Temkin Isotherm Analysis
}
\description{
takes into account the effects of indirect adsorbate/adsorbate interaction on the adsorption process
}
\examples{
temkinanalysis (c(1,2,3,4,5),c(1,2,3,4,5))
}
