% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temkinplot.R
\name{temkinplot-deprecated}
\alias{temkinplot-deprecated}
\title{Temkin Isotherm Analysis Plot-deprecated}
\usage{
temkinplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the Temkin Isotherm Analysis
}
\description{
takes into account the effects of indirect adsorbate/adsorbate interaction on the adsorption process
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
