% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temkinplot.R
\name{temkinplot}
\alias{temkinplot}
\title{Temkin Isotherm Analysis}
\usage{
temkinplot(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the plot for the Temkin Isotherm Analysis
}
\description{
takes into account the effects of indirect adsorbate/adsorbate interaction on the adsorption process
}
\examples{
temkinplot (c(1,2,3,4,5),c(1,2,3,4,5))
}
