% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tothplot.R
\name{tothplot-deprecated}
\alias{tothplot-deprecated}
\title{Toth Isotherm Plot-deprecated}
\usage{
tothplot(Ce, theta)
}
\arguments{
\item{Ce}{the numeric value for the equilibrium concentration}

\item{theta}{the numeric value for the fractional coverage}
}
\value{
the LSRL plot for a toth isotherm analysis
}
\description{
Plot of the analysis of Toth Isotherm
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
