% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tothplot.R
\name{tothplot}
\alias{tothplot}
\title{Toth Isotherm Plot}
\usage{
tothplot(Ce, theta)
}
\arguments{
\item{Ce}{the numeric value for the equilibrium concentration}

\item{theta}{the numeric value for the fractional coverage}
}
\value{
the LSRL plot for a toth isotherm analysis
}
\description{
Plot of the analysis of Toth Isotherm
}
\examples{
tothplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
