% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVVplot.R
\name{WVVplot-deprecated}
\alias{WVVplot-deprecated}
\title{Weber Van Vliet Isotherm Plot-deprecated}
\usage{
WVVplot(Qe, Ce)
}
\arguments{
\item{Qe}{the numeric value for the adsorbed concentration}

\item{Ce}{the numeric value for the equilibrium concentration}
}
\value{
the linear plot for the Weber Van Vliet Isotherm
}
\description{
Weber Van Vliet Isotherm Plot
}
\seealso{
\code{\link{PUPAIM-deprecated}}
}
\keyword{internal}
