% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVVplot.R
\name{WVVplot}
\alias{WVVplot}
\title{Weber Van Vliet Isotherm Plot}
\usage{
WVVplot(Qe, Ce)
}
\arguments{
\item{Qe}{the numeric value for the adsorbed concentration}

\item{Ce}{the numeric value for the equilibrium concentration}
}
\value{
the linear plot for the Weber Van Vliet Isotherm
}
\description{
Plot of the abalysis of Weber Van Vliet Isotherm
}
\examples{
WVVplot(c(1,2,3,4,5),c(1,2,3,4,5))
}
