% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banana.R
\docType{data}
\name{banana}
\alias{banana}
\title{Banana Adsorption Study}
\format{A data frame with 6 observations on the following 2 variables.
\describe{
\item{\code{Ce}}{a numeric vector}
\item{\code{Qe}}{a numeric vector}

}
...}
\usage{
banana
}
\description{
Actual adosrption experiment results using Musa paradisiaca peels.
}
\keyword{datasets}
