% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETanalysis.R
\name{BETanalysis}
\alias{BETanalysis}
\title{BET Isotherm Analysis Non-Linear Form}
\usage{
BETanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the non-linear regression, errors, and the parameter values for the BET Isotherm Model
}
\description{
An isotherm that takes account of the possibility that the monolayer in the Langmuir adsorption isotherm can act as a substrate for further adsorption.
}
\examples{
BETanalysis(c(.014, .063, .094, .2, .385, 1.15, 1.64), c(.072, .213, .282, .32, .325, .338, .344))
}
\references{
Shipu, S.(2014, April 30). Bet Isotherm. Retrived form https://www.slideshare.net/mobile/Sourav44Shipu/bet-isotherm
}
\author{
Geraldine N. Aliman

Cherry Rose E. Olivar

C.C. Deocaris
}
