% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elovichanalysis.R
\name{elovichanalysis}
\alias{elovichanalysis}
\title{Elovich Isotherm Analysis}
\usage{
elovichanalysis(Qe, Ce)
}
\arguments{
\item{Qe}{The numeric value for the equilibrium concentration}

\item{Ce}{the numeric value for the adsorbed concentration}
}
\value{
the regression analysis for the Elovich Isotherm analysis
}
\description{
The equation that defines this model is based on a kinetic principle which assumes that adsorption sites increase exponentially with adsorption; this implies a multilayer adsorption.
}
\examples{
elovichanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
