% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5paramanalysis.R
\name{fiveparamanalysis}
\alias{fiveparamanalysis}
\title{Five Parameter Isotherm Analysis}
\usage{
fiveparamanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the five parameter isotherm analysis
}
\description{
A five-parameter empirical model that is capable of simulating the model variations more precisely for application over a wide range of equilibrium data.
}
\examples{
fiveparamanalysis(moringa$Ce,moringa$Qe)
}
