% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fowlerguggenheimanalysis.R
\name{fowlerganalysis}
\alias{fowlerganalysis}
\title{Fowler Guggenheim Isotherm}
\usage{
fowlerganalysis(theta, Ce)
}
\arguments{
\item{theta}{the numerical value reperesenting the fractional coverage}

\item{Ce}{the numerical value fot the equilibrium capacity}
}
\value{
the linear regression and the parameters for the Fowler Guggenheim isotherm analysis
}
\description{
This isotherm equation which takes into consideration the lateral interaction of the adsorbed molecules.
}
\examples{
fowlerganalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
