% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freundlichanalysis.R
\name{freundlichanalysis}
\alias{freundlichanalysis}
\title{Freundlich Isotherm Analysis}
\usage{
freundlichanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the regression analysis for the freundlich isotherm analysis
}
\description{
Gives an expression which defines the surface heterogeneity and the exponential distribution of active sites and their energies
}
\examples{
freundlichanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
