% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fritzsclunderanalysis.R
\name{fritzanalysis}
\alias{fritzanalysis}
\title{Fritz-Schluender Isotherm Analysis Non-Linear Form}
\usage{
fritzanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the nonlinear regression and the parameters for the Fritz-Schlunder isotherm
}
\description{
An empirical equation which can fit a wide range of experimental results because of the large number of coefficients in the isotherm.
}
\examples{
fritzanalysis(moringa$Ce, moringa$Qe)
}
\references{
Aljeboree, A. M., Alshirifi, A. N., & Alkaim, A. F. (2017). Kinetics and equilibrium study for the adsorption of textile dyes on coconut shell activated carbon. Arabian Journal of Chemistry, 10, S3381-S3393. doi: 10.1016/j.arabjc.2014.01.020
}
\author{
Diego Calamba

Rain Grace R. Reyes

C.C. Deocaris
}
