% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halseyanalysis.R
\name{halseyanalysis}
\alias{halseyanalysis}
\title{Halsey Isotherm}
\usage{
halseyanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Halsey isotherm analysis
}
\description{
Used to evaluate multilayer adsorption at a relatively large distance from the surface.
}
