% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harkinsjuraanalysis.R
\name{harkinsjuraanalysis}
\alias{harkinsjuraanalysis}
\title{HarkinsJura Isotherm}
\usage{
harkinsjuraanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Harkins Jura isotherm analysis
}
\description{
It assumes the possibility of multilayer adsorption on the surface of absorbents having heterogeneous pore distribution.
}
\examples{
harkinsjuraanalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
