% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hilldeboeranalysis.R
\name{hilldeboeranalysis}
\alias{hilldeboeranalysis}
\title{Hill Deboer Isotherm Analysis}
\usage{
hilldeboeranalysis(theta, Ce)
}
\arguments{
\item{theta}{the numerical value representing the fractional coverage}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the linear regression and the parameters for the Hill Deboer isotherm analysis
}
\description{
Describes an incident where there is mobile adsorption as well as lateral interaction among adsorbed molecules.
}
\examples{
hilldeboeranalysis(c(1,2,3,4,5),c(1,2,3,4,5))
}
