\name{moringa}
\alias{moringa}
\docType{data}
\title{
Moringa Adsorption Study
}
\description{
Adsorption of Copper(II) using Moringa oleifera seeds as bioadsorbent
}
\usage{data("moringa")}

\format{
  A data frame with 6 observations on the following 2 variables.
  \describe{
    \item{Ce}{a numeric vector for the equilibrium concentration}
    \item{Qe}{a numeric vector for the adsorbed concentration}

  }
}
\source{
Development of an R package for chemical adsorption isotherms, kinetics and thermodynamics(2019), BS Chemistry thesis , Polytechnic University of the Philippines
}
