% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raudkeprausniizanalysis.R
\name{raudkepanalysis}
\alias{raudkepanalysis}
\title{Raudke Prausniiz Isotherm Analysis}
\usage{
raudkepanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Raudke Prausniiz isotherm analysis
}
\description{
It has several important properties which makes it more preferred in most adsorption systems at low adsorbate concentration.
}
\examples{
raudkepanalysis(moringa$Ce,moringa$Qe)
}
