% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redlichpetersonanalysis.R
\name{redlichpanalysis}
\alias{redlichpanalysis}
\title{Redlich-Peterson Non-Linear Form}
\usage{
redlichpanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the linear regression and the parameters for the Redlich-Peterson isotherm
}
\description{
It is used for three parameter adsorption and its a combination of Langmuir and Freundlich Isotherms
}
\examples{
redlichpanalysis(moringa$Ce,moringa$Qe)
}
\references{
Ayawei, N., Ebelegi, A. N., & Wankasi, D. (2017). Modelling and Interpretation of Adsorption Isotherms. Journal of Chemistry, 2017, 1.11. doi: 10.1155/2017/3039817

Wu, F.-C., Liu, B.-L., Wu, K.-T., & Tseng, R.-L. (2010). A new linear form analysis of Redlich?Peterson isotherm equation for the adsorptions of dyes. Chemical Engineering Journal, 162(1), 21.27. doi: 10.1016/j.cej.2010.03.006
}
\author{
John Carlo F. Panganiban

C.C. Deocaris
}
