% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryanalysis.R
\name{summaryanalysis}
\alias{summaryanalysis}
\title{Summary of the Iostherm Analysis}
\usage{
summaryanalysis(Ce, Qe)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{Qe}{the numerical value for the adsorbed capacity}
}
\value{
the summary of the linear model for different adsorption isotherm
}
\description{
Summarize the analysis for different isotherm models
}
