% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tothanalysis.R
\name{tothanalysis}
\alias{tothanalysis}
\title{Toth Isotherm}
\usage{
tothanalysis(Ce, theta)
}
\arguments{
\item{Ce}{the numerical value for the equilibrium capacity}

\item{theta}{the numerical value representing the fractional coverage}
}
\value{
the linear regression and the parameters for the Toth isotherm analysis
}
\description{
Another empirical modification of the Langmuir equation with the aim of reducing the error between experimental data and predicted value of equilibrium data.
}
