% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WVVanalysis.R
\name{webervvanalysis}
\alias{webervvanalysis}
\title{Weber Van Vliet Isotherm Analysis}
\usage{
webervvanalysis(Qe, Ce)
}
\arguments{
\item{Qe}{the numerical value for the adsorbed capacity}

\item{Ce}{the numerical value for the equilibrium capacity}
}
\value{
the linear regression and the parameters for the Weber Van Vliet isotherm analysis
}
\description{
An empirical relation with four parameters that provided excellent description of data patterns for a wide range of adsorption systems.
}
